+++ backend/llmspot/__init__.py
@@ -1,0 +1,3 @@
+from .models import openaiwrp, anthropicwrp
+
+__all__ = ['openaiwrp', 'anthropicwrp', 'base']